what ever

+++ microcontroller/src-atmel/playground/canmacht/stuff.c
@@ -27,7 +27,7 @@ AVRX_GCC_TASKDEF(stuff, 100, 4) {
 			dispFillRect(80, 30 + 9 * i, 30, 8, 0);
 			itoa(aktSensor->typ, bla, 10);
 			strcat(bla, " ");
-			itoa(aktSensor->wert, bla2, 10);
+			itoa((int) aktSensor->wert, bla2, 10);
 			strcat(bla, bla2);
 			draw_Text(bla, 80, 30 + 9 * i, 0, 1, 1);
 			aktSensor = aktSensor->next;
